requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.3